Refactor
Make CaeLoaderService smaller by moving different logic to separate
classes. This way is easier to write unit tests and maintain code.
